Move WbeExportTx/MintBuilder to Types module
